@@ -3,28 +3,26 @@
 #include <netinet/in.h>
 
 #include <iostream>
-#include <string>
 
 namespace NetUtils
 {
 
 enum class Error
 {
+	NO_ERROR,
 	PORT_REQUIRED,
 	PARSE_ERROR
 };
 
-extern Error error;
+const char *StrError(Error e);
 
-void PrintError(const std::string &prefix);
-bool ParseAddress(const char *addr, const int defaultPort, sockaddr_in &sain);
+// Parse an address in the form `host:port`.
+// Domain names will be resolved using `gethostbyname()`.
+// If `defaultPort` is `-1`, the `:port` part of `addr` is required.
+Error ParseAddress(sockaddr_in &sain, const char *addr, const int defaultPort = -1);
+Error ParseAddress(sockaddr_in &sain, const std::string &addr, const int defaultPort = -1);
 
 } // namespace NetUtils
 
 // for printing ip addresses without first converting to a string
-inline std::ostream &operator<<(std::ostream &ostr, const in_addr &ip)
-{
-	return ostr << (ip.s_addr & 0xFF) << '.' << ((ip.s_addr >> 8) & 0xFF) << '.'
-				<< ((ip.s_addr >> 16) & 0xFF) << '.'
-				<< ((ip.s_addr >> 24) & 0xFF);
-}
+std::ostream &operator<<(std::ostream &ostr, const in_addr &ip);
