@@ -14,20 +14,9 @@ void CliPrompt::printFullPrompt(bool withInput)
 	*ostr << cursor << Cursor::moveLeftOne;
 }
 
-void CliPrompt::processUntilEnterPressed()
-{
-	prepareForNextLine();
-	printFullPrompt(false);
-	do
-	{
-		processKeyboard();
-		swiWaitForVBlank();
-	} while (pmMainLoop() && !_enterPressed);
-}
-
 void CliPrompt::prepareForNextLine()
 {
-	cursorPos = flashState = flashTimer = {};
+	cursorPos = flashState = {};
 	lineHistoryItr = lineHistory.cend();
 	savedInput.clear();
 	input.clear();
@@ -39,19 +28,15 @@ void CliPrompt::flashCursor()
 	 * While `cursorPos != line.size()`, the character pointed to by the cursor
 	 * will flash between `line[cursorPos]` and `this->cursor`.
 	 */
-	static const auto FLASH_INTERVAL = 10;
 
 	if (cursorPos == input.size())
-	{
-		flashTimer = 0;
 		flashState = false;
-	}
-	else if (++flashTimer >= FLASH_INTERVAL)
+	else
 	{
-		flashTimer = 0;
-		((flashState = !flashState) ? (*ostr << cursor)
-									: (*ostr << input[cursorPos]))
-			<< Cursor::moveLeftOne;
+		// this somehow flashes it at the exact same interval as before
+		// i implemented multithreading??? flashTimer no longer needed
+		threadSleep(ticksFromHz(60));
+		*ostr << ((flashState = !flashState) ? cursor : input[cursorPos]) << Cursor::moveLeftOne;
 	}
 }
 
@@ -64,8 +49,7 @@ void CliPrompt::handleBackspace()
 		*ostr << " \b\b" << cursor << Cursor::moveLeftOne;
 	else
 		*ostr << "\b\e[0K" << input.c_str() + cursorPos
-			  << Cursor::move(
-					 Cursor::MoveDirection::LEFT, input.size() - cursorPos);
+			  << Cursor::move(Cursor::MoveDirection::LEFT, input.size() - cursorPos);
 }
 
 void CliPrompt::handleEnter()
@@ -87,10 +71,15 @@ void CliPrompt::handleLeft()
 {
 	if (input.empty() || cursorPos == 0)
 		return;
+
 	if (cursorPos == input.size())
+		// remove static cursor
 		*ostr << " \b";
-	// if (flashState)
-		// *ostr << input[cursorPos] << Cursor::moveLeftOne;
+
+	if (flashState)
+		// prevent leaving behind cursors over the input
+		*ostr << input[cursorPos] << Cursor::moveLeftOne;
+
 	*ostr << Cursor::moveLeftOne;
 	--cursorPos;
 }
@@ -106,7 +95,7 @@ void CliPrompt::handleRight()
 
 void CliPrompt::handleUp()
 {
-	if (lineHistory.empty() || lineHistoryItr - 1 < lineHistory.cbegin())
+	if (lineHistory.empty() || lineHistoryItr == lineHistory.cbegin())
 		// obviously do nothing here
 		return;
 
@@ -122,7 +111,7 @@ void CliPrompt::handleUp()
 
 void CliPrompt::handleDown()
 {
-	if (lineHistory.empty() || lineHistoryItr + 1 > lineHistory.cend())
+	if (lineHistory.empty() || lineHistoryItr == lineHistory.cend())
 		// obviously do nothing here
 		return;
 
@@ -138,7 +127,7 @@ void CliPrompt::handleDown()
 		return;
 	}
 
-	// go down one, reset cursorPos, reprint everything
+	// reset cursorPos, reprint everything
 	input = *lineHistoryItr;
 	cursorPos = input.size();
 	*ostr << "\r\e[2K" << prompt << input << cursor << Cursor::moveLeftOne;
@@ -228,9 +217,7 @@ void CliPrompt::processKeyboard()
 		{
 			input.insert(input.begin() + cursorPos, c);
 			*ostr << input.c_str() + cursorPos
-				  << Cursor::move(
-						 Cursor::MoveDirection::LEFT,
-						 input.size() - cursorPos - 1);
+				  << Cursor::move(Cursor::MoveDirection::LEFT, input.size() - cursorPos - 1);
 		}
 		if (++cursorPos == input.size())
 			*ostr << cursor << Cursor::moveLeftOne;
