@@ -1,7 +1,6 @@
 #include "Commands.hpp"
-#include "EscapeSequences.hpp"
 #include "NetUtils.hpp"
-#include "Shell.hpp"
+#include "ShellClass.hpp"
 
 #include <dswifi9.h>
 #include <nds.h>
@@ -16,7 +15,6 @@
 
 #include <filesystem>
 #include <fstream>
-#include <iostream>
 
 namespace fs = std::filesystem;
 
@@ -26,23 +24,20 @@ extern const devoptab_t dotab_stdnull;
 namespace Commands
 {
 
-void help()
+std::string Context::GetEnv(const std::string &key, const std::string &_default) const
 {
-	auto num_commands = MAP.size();
-	*Shell::out << num_commands << " commands available:\n";
-	auto itr = MAP.cbegin();
-	for (; num_commands > 1; --num_commands)
-		*Shell::out << (itr++)->first << ' ';
-	*Shell::out << itr->first << '\n';
+	if (const auto itr = env.find(key); itr != env.cend())
+		return itr->second;
+	return _default;
 }
 
-void ls()
+void ls(const Context &ctx)
 {
-	const auto &path = (Shell::args.size() == 1) ? Shell::cwd : Shell::args[1];
+	const auto &path = (ctx.args.size() == 1) ? ctx.GetEnv("PWD") : ctx.args[1];
 
 	if (!fs::exists(path))
 	{
-		*Shell::err << "\e[41mpath does not exist\e[39m\n";
+		ctx.err << "\e[41mpath does not exist\e[39m\n";
 		return;
 	}
 
@@ -50,275 +45,213 @@ void ls()
 	{
 		const auto filename = entry.path().filename().string();
 		if (entry.is_directory())
-			*Shell::out << "\e[44m" << filename << "\e[39m ";
+			ctx.out << "\e[44m" << filename << "\e[39m ";
 		else
-			*Shell::out << filename << ' ';
-		*Shell::out << '\n';
+			ctx.out << filename << ' ';
+		ctx.out << '\n';
 	}
 }
 
-void cd()
+void cd(const Context &ctx)
 {
-	if (Shell::args.size() == 1)
+	if (ctx.args.size() == 1)
 	{
-		Shell::cwd = "/";
+		ctx.shell.SetEnv("PWD", "/");
 		return;
 	}
 
-	const auto &path = Shell::args[1];
+	const auto &path = ctx.args[1];
 
 	if (!fs::exists(path))
 	{
-		*Shell::err << "\e[41mpath does not exist\e[39m\n";
+		ctx.err << "\e[41mpath does not exist\e[39m\n";
 		return;
 	}
 
-	Shell::cwd = path;
+	ctx.shell.SetEnv("PWD", path);
 }
 
-void cat()
+void cat(const Context &ctx)
 {
-	if (Shell::args.size() == 1)
+	if (ctx.args.size() == 1)
 	{
-		if (Shell::in == &std::cin)
+		if (&ctx.in == &std::cin)
 		{
 			// this gets you stuck in the cat command,
 			// unless i make stdin nonblocking and use threads
 			// so you can "ctrl+c" by pressing the fold key
-			*Shell::err << "\e[41mcat: not using stdin\e[39m\n";
+			ctx.err << "\e[41mcat: not using stdin\e[39m\n";
 			return;
 		}
 
-		*Shell::out << Shell::in->rdbuf();
+		ctx.out << ctx.in.rdbuf();
 		return;
 	}
 
 	std::error_code ec;
-	if (!fs::exists(Shell::args[1], ec) && !ec)
+	if (!fs::exists(ctx.args[1], ec) && !ec)
 	{
-		*Shell::err << "\e[41mcat: file does not exist: " << Shell::args[1] << "\e[39m\n";
+		ctx.err << "\e[41mcat: file does not exist: " << ctx.args[1] << "\e[39m\n";
 		return;
 	}
 	else if (ec)
 	{
-		*Shell::err << "\e[41mcat: " << ec.message() << "\e[39m\n";
+		ctx.err << "\e[41mcat: " << ec.message() << "\e[39m\n";
 		return;
 	}
 
-	const std::ifstream file{Shell::args[1]};
+	const std::ifstream file{ctx.args[1]};
 
 	if (!file)
 	{
-		*Shell::err << "\e[41mcat: cannot open file: " << Shell::args[1] << "\e[39m\n";
+		ctx.err << "\e[41mcat: cannot open file: " << ctx.args[1] << "\e[39m\n";
 		return;
 	}
 
-	*Shell::out << file.rdbuf();
+	ctx.out << file.rdbuf();
 }
 
-void rm()
+void rm(const Context &ctx)
 {
-	if (Shell::args.size() == 1)
+	if (ctx.args.size() == 1)
 	{
-		*Shell::err << "args: <filepaths...>\n";
+		ctx.err << "args: <filepaths...>\n";
 		return;
 	}
 
-	for (auto itr = Shell::args.cbegin() + 1; itr < Shell::args.cend(); ++itr)
+	for (auto itr = ctx.args.cbegin() + 1; itr < ctx.args.cend(); ++itr)
 	{
 		std::error_code ec;
-		if (!fs::remove(Shell::args[1], ec))
-			*Shell::err << "\e[41mrm: failed to remove: '" << Shell::args[1] << "': " << ec.message() << "\e[39m\n";
+		if (!fs::remove(ctx.args[1], ec))
+			ctx.err << "\e[41mrm: failed to remove: '" << ctx.args[1] << "': " << ec.message() << "\e[39m\n";
 	}
 }
 
-void echo()
+void echo(const Context &ctx)
 {
-	for (auto itr = Shell::args.cbegin() + 1; itr < Shell::args.cend(); ++itr)
-		*Shell::out << *itr << ' ';
-	*Shell::out << '\n';
+	for (auto itr = ctx.args.cbegin() + 1; itr < ctx.args.cend(); ++itr)
+		ctx.out << *itr << ' ';
+	ctx.out << '\n';
 }
 
-void dns()
+void dns(const Context &ctx)
 {
-	if (Shell::args.size() == 1)
+	if (ctx.args.size() == 1)
 	{
-		*Shell::err << "args: <hostname>\n";
+		ctx.err << "args: <hostname>\n";
 		return;
 	}
 
-	const auto host = gethostbyname(Shell::args[1].c_str());
+	const auto host = gethostbyname(ctx.args[1].c_str());
 	if (!host)
 	{
-		*Shell::err << "\e[41mdns: gethostbyname: " << strerror(errno) << "\e[39m";
+		ctx.err << "\e[41mdns: gethostbyname: " << strerror(errno) << "\e[39m";
 		return;
 	}
 
-	*Shell::out << *(in_addr *)host->h_addr_list[0] << '\n';
+	ctx.out << *(in_addr *)host->h_addr_list[0] << '\n';
 }
 
-void ip()
+void ip(const Context &ctx)
 {
-	*Shell::out << (in_addr)Wifi_GetIP() << '\n';
+	ctx.out << (in_addr)Wifi_GetIP() << '\n';
 }
 
-void ipinfo()
+void ipinfo(const Context &ctx)
 {
 	in_addr gateway, subnetMask, dns1, dns2;
 	Wifi_GetIPInfo(&gateway, &subnetMask, &dns1, &dns2);
 
-	*Shell::out << "IP:          " << (in_addr)Wifi_GetIP() << '\n'
-				<< "Gateway:     " << gateway << '\n'
-				<< "Subnet Mask: " << subnetMask << '\n'
-				<< "DNS 1:       " << dns1 << '\n'
-				<< "DNS 2:       " << dns2 << '\n';
+	ctx.out << "IP:          " << (in_addr)Wifi_GetIP() << '\n'
+			<< "Gateway:     " << gateway << '\n'
+			<< "Subnet Mask: " << subnetMask << '\n'
+			<< "DNS 1:       " << dns1 << '\n'
+			<< "DNS 2:       " << dns2 << '\n';
 }
 
-void source()
+void source(const Context &ctx)
 {
-	if (Shell::args.size() < 2)
+	if (ctx.args.size() < 2)
 	{
-		*Shell::err << "args: <filepath>\n";
+		ctx.err << "args: <filepath>\n";
 		return;
 	}
 
-	Shell::SourceFile(Shell::args[1]);
+	ctx.shell.SourceFile(ctx.args[1]);
 }
 
-void rename()
+void rename(const Context &ctx)
 {
-	if (Shell::args.size() < 3)
+	if (ctx.args.size() < 3)
 	{
-		*Shell::err << "args: <oldpath> <newpath>\n";
+		ctx.err << "args: <oldpath> <newpath>\n";
 		return;
 	}
 
 	std::error_code ec;
-	fs::rename(Shell::args[1], Shell::args[2], ec);
+	fs::rename(ctx.args[1], ctx.args[2], ec);
 	if (ec)
-		*Shell::err << "\e[41mrename: " << ec.message() << "\e[39m\n";
+		ctx.err << "\e[41mrename: " << ec.message() << "\e[39m\n";
 }
 
-void pwd()
+void pwd(const Context &ctx)
 {
-	*Shell::out << Shell::cwd << '\n';
+	ctx.out << ctx.GetEnv("PWD") << '\n';
 }
 
-void env()
+void env(const Context &ctx)
 {
-	for (const auto &[k, v] : Shell::env)
-		*Shell::out << k << '=' << v << '\n';
+	for (const auto &[k, v] : ctx.env)
+		ctx.out << k << '=' << v << '\n';
 }
 
-void unset()
+void unset(const Context &ctx)
 {
-	if (Shell::args.size() < 2)
+	if (ctx.args.size() < 2)
 	{
-		*Shell::err << "args: <envkey>\n";
+		ctx.err << "args: <envkey>\n";
 		return;
 	}
 
-	Shell::env.erase(Shell::args[1]);
+	ctx.shell.UnsetEnv(ctx.args[1]);
 }
 
-// void history()
-// {
-// 	if (Shell::args.size() == 2 && Shell::args[1] == "-c")
-// 	{
-// 		Shell::prompt.clearLineHistory();
-// 		std::error_code ec;
-// 		if (!fs::remove("/.ndsh_history", ec))
-// 			*Shell::err << "\e[41mhistory: failed to remove '.ndsh_history': " << ec.message() << "\e[39m\n";
-// 		return;
-// 	}
+void history(const Context &ctx)
+{
+	if (ctx.args.size() == 2 && ctx.args[1] == "-c")
+	{
+		ctx.shell.ClearLineHistory();
+		std::error_code ec;
+		if (!fs::remove("/.ndsh_history", ec))
+			ctx.err << "\e[41mhistory: failed to remove '.ndsh_history': " << ec.message() << "\e[39m\n";
+		return;
+	}
 
-// 	for (const auto &line : Shell::prompt.getLineHistory())
-// 		*Shell::out << line << '\n';
-// }
+	for (const auto &line : ctx.shell.GetLineHistory())
+		ctx.out << line << '\n';
+}
 
-void devices()
+void devices(const Context &ctx)
 {
 	for (int i = 0; i < STD_MAX; ++i)
 	{
 		const auto dev = devoptab_list[i];
 		if (!dev || dev == &dotab_stdnull)
 			continue;
-		*Shell::out << i << ':' << dev->name << ' ';
+		ctx.out << i << ':' << dev->name << ' ';
 	}
-	*Shell::out << '\n';
+	ctx.out << '\n';
 }
 
-/*void consoles_test()
+void exit(const Context &ctx)
 {
-	consoleClear();
+	// Exit the current shell.
+	ctx.shell.SetShouldExit();
+}
 
-	std::ofstream con_streams[Shell::NUM_CONSOLES];
-	CliPrompt *prompts[Shell::NUM_CONSOLES];
-	std::string lines[Shell::NUM_CONSOLES];
+void help(const Context &);
 
-	for (int i = 0; i < Shell::NUM_CONSOLES; ++i)
-	{
-		prompts[i] = new CliPrompt{"con" + std::to_string(i) + "> ", '_', con_streams[i]};
-
-		auto con_name = "con" + std::to_string(i) + ':';
-		con_streams[i].rdbuf()->pubsetbuf(nullptr, 0);
-		con_streams[i].open(con_name);
-		if (!con_streams[i])
-			*Shell::err << "failed to open con" << i << '\n';
-		con_streams[i] << "hello console " << i << "!\npress start to exit\n"
-					   << prompts[i]->prompt << prompts[i]->cursor << EscapeSequences::Cursor::moveLeftOne;
-	}
-
-	int focused_console{};
-
-	while (pmMainLoop())
-	{
-		swiWaitForVBlank();
-		scanKeys();
-
-		const auto keys = keysDown();
-
-		if (keys & KEY_START)
-			break;
-
-		if ((keys & KEY_L) && focused_console > 0)
-		{
-			bgHide(Shell::consoles[focused_console].bgId);
-			bgShow(Shell::consoles[--focused_console].bgId);
-		}
-
-		if ((keys & KEY_R) && focused_console < Shell::NUM_CONSOLES - 1)
-		{
-			bgHide(Shell::consoles[focused_console].bgId);
-			bgShow(Shell::consoles[++focused_console].bgId);
-		}
-
-		auto prompt = prompts[focused_console];
-		auto &con_stream = con_streams[focused_console];
-
-		prompt->processKeyboard(lines[focused_console]);
-
-		if (prompt->newlineEntered())
-			con_stream << prompt->prompt << prompt->cursor << EscapeSequences::Cursor::moveLeftOne;
-	}
-
-	// switch back to console 0 on main display
-	if (focused_console > 0 && focused_console < 4)
-	{
-		bgHide(Shell::consoles[focused_console].bgId);
-		bgShow(Shell::consoles[0].bgId);
-	}
-
-	// clear all bottom consoles
-	for (int i = 4; i < Shell::NUM_CONSOLES; ++i)
-		con_streams[i] << "\e[2J";
-
-	// free our custom prompts
-	for (int i = 0; i < Shell::NUM_CONSOLES; ++i)
-		delete prompts[i];
-}*/
-
-const std::unordered_map<std::string, void (*)()> MAP{
+const Map MAP{
 	{"help", help},
 	{"echo", echo},
 	{"cd", cd},
@@ -334,16 +267,26 @@ const std::unordered_map<std::string, void (*)()> MAP{
 	{"pwd", pwd},
 	{"env", env},
 	{"rename", rename},
-	{"clear", consoleClear},
+	{"clear",
+	 [](auto &ctx)
+	 {
+		 ctx.out << "\e[2J";
+	 }},
 	{"unset", unset},
-	// {"history", history},
+	{"history", history},
 	{"devices", devices},
-	// {"contest", consoles_test},
-	// this is better than calling exit(), as it lets us perform cleanup
-	// before our code fully returns. the power button is a hard-reset
-	// unfortunately.
-	{"exit", pmPrepareToReset},
+	{"exit", exit},
 	{"lua", lua},
 	{"source", source}};
+
+void help(const Context &ctx)
+{
+	auto num_commands = MAP.size();
+	ctx.out << num_commands << " commands available:\n";
+	auto itr = MAP.cbegin();
+	for (; num_commands > 1; --num_commands)
+		ctx.out << (itr++)->first << ' ';
+	ctx.out << itr->first << '\n';
+}
 
 } // namespace Commands
